move rdf_page_to_resource.rb and rdf_resolve_prefix.rb to lib/jekyll/helper/
